backend/requirements.txt
# Backend Requirements for AI Emotion Monitor Flask API

# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0

# Flask API
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
python-dotenv>=1.0.0
